# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CK Empire Builder API
  description: |
    # CK Empire Builder - Advanced Digital Empire Management API
    
    ## Overview
    
    CK Empire Builder is a comprehensive digital empire management platform that provides:
    
    * **Project Management**: Organize and track digital projects
    * **Revenue Tracking**: Monitor income streams and financial metrics
    * **AI Integration**: Advanced AI features for content generation
    * **Ethics Monitoring**: Built-in ethical AI monitoring
    * **Performance Analytics**: Real-time metrics and insights
    * **Cloud Integration**: AWS/GCP integration with auto-scaling
    * **Backup & Recovery**: Automated backup and disaster recovery
    
    ## Features
    
    ### 🔥 Core Features
    - **Project Management**: Create, track, and manage digital projects
    - **Revenue Analytics**: Monitor income streams and financial performance
    - **AI-Powered Content**: Generate content using advanced AI models
    - **Ethics Compliance**: Built-in ethical AI monitoring and compliance
    - **Performance Monitoring**: Real-time analytics and insights
    
    ### ☁️ Cloud Features
    - **AWS Integration**: S3 backup, RDS database, ElastiCache Redis
    - **GCP Integration**: Cloud Storage, Cloud SQL, Memorystore
    - **Auto-scaling**: Kubernetes HPA for automatic scaling
    - **Backup & Recovery**: Automated database backups and restoration
    - **Monitoring**: CloudWatch metrics and alerts
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/ai/ideas:
    post:
      summary: Generate AI Ideas
      description: Generate AI-powered ideas for content or projects
      tags:
        - AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdeaGenerationRequest'
      responses:
        '200':
          description: Ideas generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdeaGenerationResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    IdeaGenerationRequest:
      type: object
      required:
        - topic
      properties:
        topic:
          type: string
          description: Topic for idea generation
          example: "artificial intelligence"
        style:
          type: string
          enum: [viral, professional, creative, technical]
          default: viral
          example: "viral"
        count:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          example: 5
    IdeaGenerationResponse:
      type: object
      properties:
        ideas:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                example: "AI-Powered Personal Assistant"
              description:
                type: string
                example: "Create an AI assistant that learns user preferences"
              category:
                type: string
                example: "Technology"
              potential_revenue:
                type: number
                format: float
                example: 50000.00
        total_generated:
          type: integer
          example: 5
...