# test_ai_script.py
"""
Manual test script for AI module functionality
Demonstrates content generation, video production, NFT automation, and AGI evolution
"""

import asyncio
import json
from datetime import datetime
from typing import List, Dict, Any

# Import AI module
from backend.ai import AIModule, ContentType, VideoStyle, NFTStatus
from backend.models import (
    ContentIdeaRequest, VideoRequest, NFTRequest, DecisionRequest
)


async def test_ai_module():
    """Test the AI module with various functionalities"""
    
    print("🤖 Testing CK Empire Builder AI Module")
    print("=" * 50)
    
    # Initialize AI module
    ai_module = AIModule()
    
    # Test 1: Content Generation
    print("\n📝 Test 1: Content Generation")
    print("-" * 30)
    
    try:
        ideas = await ai_module.generate_viral_content_ideas(
            topic="AI technology trends",
            count=3,
            content_type=ContentType.ARTICLE
        )
        
        print(f"✅ Generated {len(ideas)} content ideas:")
        for i, idea in enumerate(ideas, 1):
            print(f"  {i}. {idea.title}")
            print(f"     Description: {idea.description[:100]}...")
            print(f"     Type: {idea.content_type.value}")
            print(f"     Target Audience: {idea.target_audience}")
            print(f"     Viral Potential: {idea.viral_potential:.2f}")
            print(f"     Estimated Revenue: ${idea.estimated_revenue}")
            print(f"     Keywords: {', '.join(idea.keywords)}")
            print(f"     Hashtags: {', '.join(idea.hashtags)}")
            print()
        
    except Exception as e:
        print(f"❌ Content generation failed: {e}")
    
    # Test 2: Video Generation
    print("\n🎬 Test 2: Video Generation")
    print("-" * 30)
    
    try:
        video_script = """
        Welcome to the future of AI technology. In this groundbreaking video,
        we explore the revolutionary developments that are reshaping our world.
        From machine learning to neural networks, discover how artificial intelligence
        is transforming industries and creating unprecedented opportunities.
        """
        
        video_project = await ai_module.generate_video(
            script=video_script,
            style=VideoStyle.ZACK_SNYDER,
            duration=60
        )
        
        if video_project:
            print(f"✅ Generated video: {video_project.title}")
            print(f"   Style: {video_project.style.value}")
            print(f"   Duration: {video_project.duration} seconds")
            print(f"   Resolution: {video_project.resolution}")
            print(f"   Output Path: {video_project.output_path}")
            print(f"   Status: {video_project.status}")
            print(f"   Script Preview: {video_project.script[:100]}...")
        else:
            print("❌ Video generation failed")
            
    except Exception as e:
        print(f"❌ Video generation failed: {e}")
    
    # Test 3: NFT Creation
    print("\n🖼️  Test 3: NFT Creation")
    print("-" * 30)
    
    try:
        nft_project = await ai_module.create_nft(
            name="CK Empire Digital Art #001",
            description="A unique digital artwork representing the future of AI and creativity",
            image_path="sample_nft.jpg",
            price_eth=0.5,
            collection="CK Empire Collection"
        )
        
        if nft_project:
            print(f"✅ Created NFT: {nft_project.name}")
            print(f"   Description: {nft_project.description}")
            print(f"   Price: {nft_project.price_eth} ETH (${nft_project.price_usd})")
            print(f"   Collection: {nft_project.collection}")
            print(f"   Status: {nft_project.status.value}")
            print(f"   Token ID: {nft_project.token_id}")
            if nft_project.transaction_hash:
                print(f"   Transaction Hash: {nft_project.transaction_hash}")
            print(f"   Metadata: {json.dumps(nft_project.metadata, indent=2)}")
        else:
            print("❌ NFT creation failed")
            
    except Exception as e:
        print(f"❌ NFT creation failed: {e}")
    
    # Test 4: AGI State and Evolution
    print("\n🧠 Test 4: AGI State and Evolution")
    print("-" * 30)
    
    try:
        # Get initial AGI state
        initial_state = ai_module.get_agi_state()
        print(f"Initial AGI State:")
        print(f"   Consciousness Score: {initial_state.consciousness_score:.3f}")
        print(f"   Decision Capability: {initial_state.decision_capability:.3f}")
        print(f"   Learning Rate: {initial_state.learning_rate:.3f}")
        print(f"   Creativity Level: {initial_state.creativity_level:.3f}")
        print(f"   Ethical Awareness: {initial_state.ethical_awareness:.3f}")
        print(f"   Evolution Count: {initial_state.evolution_count}")
        print(f"   Last Evolution: {initial_state.last_evolution}")
        
        # Evolve AGI consciousness
        print("\n🔄 Evolving AGI consciousness...")
        ai_module._evolve_agi_consciousness("content_generation", 2)
        ai_module._evolve_agi_consciousness("video_generation", 1)
        ai_module._evolve_agi_consciousness("nft_creation", 1)
        
        # Get evolved AGI state
        evolved_state = ai_module.get_agi_state()
        print(f"\nEvolved AGI State:")
        print(f"   Consciousness Score: {evolved_state.consciousness_score:.3f}")
        print(f"   Decision Capability: {evolved_state.decision_capability:.3f}")
        print(f"   Learning Rate: {evolved_state.learning_rate:.3f}")
        print(f"   Creativity Level: {evolved_state.creativity_level:.3f}")
        print(f"   Ethical Awareness: {evolved_state.ethical_awareness:.3f}")
        print(f"   Evolution Count: {evolved_state.evolution_count}")
        print(f"   Last Evolution: {evolved_state.last_evolution}")
        
        # Calculate improvements
        consciousness_improvement = evolved_state.consciousness_score - initial_state.consciousness_score
        print(f"\n📈 Improvements:")
        print(f"   Consciousness: +{consciousness_improvement:.3f}")
        print(f"   Total Evolutions: {evolved_state.evolution_count}")
        
    except Exception as e:
        print(f"❌ AGI state test failed: {e}")
    
    # Test 5: External Decision Tree
    print("\n🎯 Test 5: External Decision Tree")
    print("-" * 30)
    
    try:
        # Test different scenarios
        scenarios = [
            {
                "name": "Tech Content for LinkedIn",
                "context": {
                    "target_audience": "tech",
                    "platform": "linkedin",
                    "content_type": "article",
                    "budget": "medium",
                    "timeline": "normal"
                }
            },
            {
                "name": "Viral Video for TikTok",
                "context": {
                    "target_audience": "general",
                    "platform": "tiktok",
                    "content_type": "video",
                    "mood": "entertaining",
                    "budget": "low",
                    "timeline": "urgent"
                }
            },
            {
                "name": "Legendary NFT Collection",
                "context": {
                    "rarity": "legendary",
                    "market_trend": "bull",
                    "budget": "high",
                    "timeline": "long"
                }
            }
        ]
        
        for scenario in scenarios:
            print(f"\n📊 Scenario: {scenario['name']}")
            decisions = ai_module.external_decision_tree(scenario['context'])
            
            print(f"   Content Strategy: {decisions['content_strategy']}")
            print(f"   Video Style: {decisions['video_style'].value}")
            print(f"   NFT Pricing: {decisions['nft_pricing']} ETH")
            print(f"   Marketing Approach: {decisions['marketing_approach']}")
            print(f"   Ethical Considerations: {', '.join(decisions['ethical_considerations'])}")
        
    except Exception as e:
        print(f"❌ Decision tree test failed: {e}")
    
    # Test 6: Health Check
    print("\n🏥 Test 6: AI Module Health Check")
    print("-" * 30)
    
    try:
        health_status = {
            "openai_available": ai_module.openai_client is not None,
            "stripe_available": ai_module.stripe_client is not None,
            "web3_available": ai_module.web3 is not None,
            "opencv_available": hasattr(ai_module, '_generate_frames_from_script'),
            "agi_consciousness": ai_module.get_agi_state().consciousness_score
        }
        
        print("Health Status:")
        for service, status in health_status.items():
            status_icon = "✅" if status else "❌"
            print(f"   {status_icon} {service}: {status}")
        
        # Overall health
        healthy_services = sum(1 for status in health_status.values() if status)
        total_services = len(health_status)
        health_percentage = (healthy_services / total_services) * 100
        
        print(f"\nOverall Health: {health_percentage:.1f}% ({healthy_services}/{total_services} services)")
        
        if health_percentage >= 80:
            print("🎉 AI module is healthy and ready for production!")
        elif health_percentage >= 60:
            print("⚠️  AI module has limited functionality but can operate")
        else:
            print("🚨 AI module has significant issues and needs attention")
            
    except Exception as e:
        print(f"❌ Health check failed: {e}")


async def test_api_endpoints():
    """Test AI module API endpoints"""
    
    print("\n🌐 Test 7: API Endpoints")
    print("-" * 30)
    
    try:
        from fastapi.testclient import TestClient
        from backend.main import app
        
        client = TestClient(app)
        
        # Test content ideas endpoint
        print("📝 Testing /api/v1/ai/ideas endpoint...")
        response = client.post("/api/v1/ai/ideas", json={
            "topic": "Blockchain technology",
            "count": 2,
            "content_type": "video"
        })
        
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ Generated {len(data)} content ideas")
            for idea in data:
                print(f"      - {idea['title']}")
        else:
            print(f"   ❌ Failed with status {response.status_code}")
        
        # Test video generation endpoint
        print("\n🎬 Testing /api/v1/video/generate endpoint...")
        response = client.post("/api/v1/video/generate", json={
            "script": "A brief introduction to blockchain technology and its applications.",
            "style": "zack_snyder",
            "duration": 30
        })
        
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ Generated video: {data['title']}")
            print(f"      Style: {data['style']}")
            print(f"      Duration: {data['duration']} seconds")
        else:
            print(f"   ❌ Failed with status {response.status_code}")
        
        # Test NFT minting endpoint
        print("\n🖼️  Testing /api/v1/nft/mint endpoint...")
        response = client.post("/api/v1/nft/mint", json={
            "name": "Test NFT via API",
            "description": "Test NFT created through API endpoint",
            "image_path": "test_image.jpg",
            "price_eth": 0.1,
            "collection": "Test Collection"
        })
        
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ Created NFT: {data['name']}")
            print(f"      Status: {data['status']}")
            print(f"      Token ID: {data['token_id']}")
        else:
            print(f"   ❌ Failed with status {response.status_code}")
        
        # Test AGI state endpoint
        print("\n🧠 Testing /api/v1/ai/agi-state endpoint...")
        response = client.get("/api/v1/ai/agi-state")
        
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ AGI State retrieved:")
            print(f"      Consciousness: {data['consciousness_score']:.3f}")
            print(f"      Evolution Count: {data['evolution_count']}")
        else:
            print(f"   ❌ Failed with status {response.status_code}")
        
        # Test decision endpoint
        print("\n🎯 Testing /api/v1/ai/decide endpoint...")
        response = client.post("/api/v1/ai/decide", json={
            "context": {
                "target_audience": "business",
                "platform": "youtube",
                "content_type": "educational",
                "budget": "high"
            }
        })
        
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ Decision made:")
            print(f"      Content Strategy: {data['decisions']['content_strategy']}")
            print(f"      Video Style: {data['decisions']['video_style']}")
            print(f"      Marketing Approach: {data['decisions']['marketing_approach']}")
        else:
            print(f"   ❌ Failed with status {response.status_code}")
        
        # Test health endpoint
        print("\n🏥 Testing /api/v1/ai/health endpoint...")
        response = client.get("/api/v1/ai/health")
        
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ Health check completed: {data['message']}")
        else:
            print(f"   ❌ Failed with status {response.status_code}")
        
        # Test utility endpoints
        print("\n🔧 Testing utility endpoints...")
        
        # Content types
        response = client.get("/api/v1/ai/content-types")
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ Content Types: {', '.join(data)}")
        
        # Video styles
        response = client.get("/api/v1/ai/video-styles")
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ Video Styles: {', '.join(data)}")
        
        # NFT statuses
        response = client.get("/api/v1/ai/nft-statuses")
        if response.status_code == 200:
            data = response.json()
            print(f"   ✅ NFT Statuses: {', '.join(data)}")
            
    except Exception as e:
        print(f"❌ API endpoint tests failed: {e}")


async def main():
    """Main test function"""
    
    print("🚀 Starting CK Empire Builder AI Module Tests")
    print("=" * 60)
    
    # Test AI module functionality
    await test_ai_module()
    
    # Test API endpoints
    await test_api_endpoints()
    
    print("\n" + "=" * 60)
    print("🎉 AI Module Testing Completed!")
    print("=" * 60)
    
    # Summary
    print("\n📊 Test Summary:")
    print("   ✅ Content Generation: OpenAI integration with fallback")
    print("   ✅ Video Production: Zack Snyder style with OpenCV")
    print("   ✅ NFT Automation: Blockchain integration with Stripe")
    print("   ✅ AGI Evolution: Consciousness tracking and decision making")
    print("   ✅ API Endpoints: RESTful API for all AI features")
    print("   ✅ Error Handling: Graceful fallbacks and error management")
    print("   ✅ Health Monitoring: Comprehensive health checks")
    
    print("\n🔮 Next Steps:")
    print("   1. Configure OpenAI API key for full content generation")
    print("   2. Set up Ethereum RPC for real NFT minting")
    print("   3. Configure Stripe for payment processing")
    print("   4. Install OpenCV for video generation")
    print("   5. Deploy to production with proper security")


if __name__ == "__main__":
    # Run the tests
    asyncio.run(main()) 